+++ mirrord/intproxy/src/request_queue.rs
@@ -8,14 +8,14 @@
 //! on this behavior and stores [`MessageId`]s of layer's requests in multiple queues. Upon
 //! receiving a response from the agent, correct [`MessageId`] is taken from the right queue.
 //!
-//! Additionaly, single [`ProxySession`](crate::session::ProxySession) handles multiple layer
+//! Additionaly, single internal proxy handles multiple layer
 //! instances (coming from forks). This fifo stores their [`SessionId`]s as well.
 
 use std::{collections::VecDeque, fmt};
 
 use thiserror::Error;
 
-use crate::protocol::{MessageId, SessionId};
+use crate::protocol::{LayerId, MessageId};
 
 /// Erorr returned when the proxy attempts to retrieve [`MessageId`] and [`SessionId`] of a request
 /// corresponding to a response received from the agent, but the [`RequestQueue`] is empty. This
@@ -29,7 +29,7 @@ pub struct RequestQueueEmpty;
 /// between them.
 #[derive(Default)]
 pub struct RequestQueue {
-    inner: VecDeque<(MessageId, SessionId)>,
+    inner: VecDeque<(MessageId, LayerId)>,
 }
 
 impl fmt::Debug for RequestQueue {
@@ -45,13 +45,13 @@ impl fmt::Debug for RequestQueue {
 impl RequestQueue {
     /// Save the request at the end of this queue.
     #[tracing::instrument(level = "trace")]
-    pub fn insert(&mut self, message_id: MessageId, session_id: SessionId) {
+    pub fn insert(&mut self, message_id: MessageId, session_id: LayerId) {
         self.inner.push_back((message_id, session_id));
     }
 
     /// Retrieve and remove a request from the front of this queue.
     #[tracing::instrument(level = "trace")]
-    pub fn get(&mut self) -> Result<(MessageId, SessionId), RequestQueueEmpty> {
+    pub fn get(&mut self) -> Result<(MessageId, LayerId), RequestQueueEmpty> {
         self.inner.pop_front().ok_or(RequestQueueEmpty)
     }
 }
